pyproject.toml
[tool.poetry]
name = "parse-large-file"
version = "0.1.0"
description = ""
authors = ["Radu Cancel <radu.cancel@gmail.com>"]
readme = "README.md"
packages = [{include = "parse_large_file"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
celery = "^5.2.7"
unicodecsv = "^0.14.1"
progressbar = "^2.5"
mysqlclient = "^2.1.1"
pandas = "^1.5.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
